requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.8
django-bootstrap-form==3.4
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-filter==2.1.0
django-heroku==0.3.1
django-session-security==2.6.5
django-tables2==2.0.6
future==0.18.2
h11==0.8.1
h2==3.1.1
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.8
image==1.5.27
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysqlclient==1.4.2
Paperboy==1.0.1
Pillow==6.0.0
psycopg2==2.8.2
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
python-dateutil==2.8.1
pytz==2019.2
requests==2.22.0
requests-async==0.6.2
rfc3986==1.3.2
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
typed-ast==1.3.5
urllib3==1.25.2
whitenoise==4.1.2
wrapt==1.11.1
